Changed EventService to EventRepository
