+++ src/main/java/se/knowit/bookitevent/configuration/BookitSpringConfiguration.java
@@ -4,15 +4,15 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import se.knowit.bookitevent.dto.TimeSupport;
 import se.knowit.bookitevent.dto.TimeSupportImpl;
-import se.knowit.bookitevent.service.EventService;
-import se.knowit.bookitevent.service.map.EventServiceMapImpl;
+import se.knowit.bookitevent.repository.EventRepository;
+import se.knowit.bookitevent.repository.EventRepositoryMapImpl;
 
 @Configuration
 public class BookitSpringConfiguration {
 
     @Bean
-    public EventService mapBasedServiceImplementation() {
-        return new EventServiceMapImpl();
+    public EventRepository mapBasedServiceImplementation() {
+        return new EventRepositoryMapImpl();
     }
 
     @Bean
