+++ src/main/java/se/knowit/bookitevent/repository/EventRepository.java
@@ -1,13 +1,17 @@
 package se.knowit.bookitevent.repository;
 
-import java.util.Optional;
-import java.util.UUID;
-
-import org.springframework.data.repository.CrudRepository;
-
 import se.knowit.bookitevent.model.Event;
 
-public interface EventRepository extends CrudRepository<Event, Long>{
+import java.util.Optional;
+import java.util.Set;
+import java.util.UUID;
+
+public interface EventRepository {
+	Set<Event> findAll();
 	
 	Optional<Event> findByEventId(UUID id);
+	
+	Optional<Event> findById(Long aLong);
+	
+	Event save(Event object);
 }
